+++ fennel-mode.el
@@ -58,20 +58,49 @@
   :type 'string
   :package-version '(fennel-mode "0.3.0"))
 
-(defcustom fennel-mode-annotate-completion t
-  "Whether or not to show kind of completion candidates."
-  :group 'fennel-mode
-  :type 'boolean
-  :package-version '(fennel-mode "0.4.0"))
+;; TODO: remove all obsolete variables in the 1.0.0 release
+(define-obsolete-variable-alias
+  'fennel-mode-annotate-completion
+  'fennel-proto-repl-annotate-completion
+  "fennel-mode 0.9.3"
+  "Completion annotation is now a feature of `fennel-proto-repl' mode.
+This variable will be removed in the 1.0.0 release of `fennel-mode'.")
 
-(defvaralias 'fennel-mode-repl-prompt-regexp
-  'fennel-mode-repl-prompt)
+(define-obsolete-variable-alias
+  'fennel-eldoc-fontify-markdown
+  'fennel-proto-repl-eldoc-fontify-markdown
+  "fennel-mode 0.7.0"
+  "Eldoc is now a feature of `fennel-proto-repl' mode.
+This variable will be removed in the 1.0.0 release of `fennel-mode'.")
 
-(make-obsolete-variable
- 'fennel-mode-repl-prompt-regexp
- 'fennel-mode-repl-prompt
- "fennel-mode 0.8.0"
- 'set)
+(define-obsolete-variable-alias
+  'fennel-mode-repl-prompt-regexp
+  'fennel-mode-repl-prompt
+  "fennel-mode 0.8.0"
+  "This variable will be removed in the 1.0.0 release of `fennel-mode'.")
+
+(define-obsolete-variable-alias
+  'fennel-mode-repl-subprompt-regexp
+  'fennel-mode-repl-subprompt
+  "fennel-mode 0.8.0"
+  "This variable will be removed in the 1.0.0 release of `fennel-mode'.")
+
+(define-obsolete-variable-alias
+  'fennel-mode-repl-log-communications
+  'fennel-proto-repl-log-communication
+  "fennel-mode 0.8.0"
+  "Logging is now a feature of `fennel-proto-repl' mode.
+This variable will be removed in the 1.0.0 release of `fennel-mode'.")
+
+(define-obsolete-variable-alias
+  'fennel-repl-minify-code
+  nil
+  "fennel-mode 0.7.0"
+  "Minification is no longer preformed; variable value is ignored.
+If you exerience problems with sending expressions to the REPL, try the
+`fennel-proto-repl' mode.
+
+This variable will be removed in the 1.0.0 release of `fennel-mode'.")
 
 (defcustom fennel-mode-repl-prompt ">>"
   "String that matches REPL prompt.
@@ -82,15 +111,6 @@ expression to match both prompt and subpromt."
   :set (lambda (sym val)
          (set-default-toplevel-value sym (string-trim-right val)))
   :package-version '(fennel-mode "0.8.0"))
-
-(defvaralias 'fennel-mode-repl-subprompt-regexp
-  'fennel-mode-repl-subprompt)
-
-(make-obsolete-variable
- 'fennel-mode-repl-subprompt-regexp
- 'fennel-mode-repl-subprompt
- "fennel-mode 0.8.0"
- 'set)
 
 (defcustom fennel-mode-repl-subprompt ".."
   "String that matches REPL subprompt.
@@ -107,17 +127,6 @@ expression to match both prompt and subpromt."
   :group 'fennel-mode
   :type 'boolean
   :package-version '(fennel-mode "0.6.0"))
-
-(defcustom fennel-mode-repl-log-communications nil
-  "Whether to log REPL communication to a hidden *fennel-repl-log* buffer."
-  :group 'fennel-mode
-  :type 'boolean
-  :package-version '(fennel-mode "0.6.0"))
-
-(make-obsolete-variable
- 'fennel-mode-repl-log-communications
- "logging is now a feature of fennel-proto-repl."
- "fennel-mode 0.8.0")
 
 (defcustom fennel-view-compilation-mode
   (cond ((fboundp 'lua-mode) #'lua-mode)
@@ -431,31 +440,6 @@ buffer, or when given a prefix arg."
   (when fennel-mode-switch-to-repl-after-reload
     (switch-to-lisp t)))
 
-(defun fennel-completion--candidate-kind (item)
-  "Annotate completion kind of the ITEM for company-mode."
-  (when fennel-mode-annotate-completion
-    (let ((type (when-let* ((buf (fennel-repl-redirect-one
-                                  (inferior-lisp-proc)
-                                  (format "(print (type %s))" item)
-                                  " *fennel-completion-annotation*")))
-                  (with-current-buffer buf
-                    (let ((contents (buffer-substring-no-properties (point-min) (point-max))))
-                      (string-trim (ansi-color-apply contents)))))))
-      (cond ((string= type "function") 'function)
-            ((string= type "table") 'module)
-            ((and type (string-match-p "tried to reference a macro" type)) 'macro)
-            ((and type (string-match-p "tried to reference a special form" type)) 'keyword)
-            ((string-match-p "\\." item) 'field)
-            (t 'variable)))))
-
-(defun fennel-completion--annotate (item)
-  "Annotate completion kind of the ITEM."
-  (let ((kind (fennel-completion--candidate-kind item)))
-    (cond ((eq kind 'module) " table")
-          ((eq kind 'variable) " definition")
-          (kind (format " %s" kind))
-          (t ""))))
-
 (defun fennel-completions (input)
   "Query completions for the INPUT from the `inferior-lisp-proc'."
   (condition-case nil
@@ -480,9 +464,7 @@ Requires Fennel 0.9.3+."
     (when completions
       (list (car bounds)
             (cdr bounds)
-            completions
-            :annotation-function #'fennel-completion--annotate
-            :company-kind #'fennel-completion--candidate-kind))))
+            completions))))
 
 (defun fennel-find-definition-for (identifier)
   "Find the definition of IDENTIFIER."
@@ -656,35 +638,6 @@ variable."
 (defvar fennel-repl--buffer-name "*Fennel REPL*")
 (defvar fennel-repl--buffer nil)
 (defvar fennel-repl--last-buffer nil)
-
-(defcustom fennel-repl-minify-code nil
-  "How to minify code before sending it to the process.
-
-The default value of `oneline' should not be changed unless there
-are some misbehaviors when interacting with the REPL.
-
-Possible variants:
-
-- \\'strings - collapse strings with literal newlines into
-  one-line strings with `\\n'.
-- \\'comments - remove comments, and lines completely consisting
-  out of comments.
-- \\'both - collapse strings and remove comments.
-- \\'oneline - collapse strings, remove comments, and remove any
-  newlines from the code.
-- nil - don't modify anything."
-  :group 'fennel-mode
-  :type '(choice
-	  (const :tag "collapse strings" strings)
-	  (const :tag "remove comments" comments)
-	  (const :tag "collapse strings and remove comments" both)
-          (const :tag "collapse to one line" oneline)
-          (const :tag "do not change" nil)))
-
-(make-obsolete-variable
- 'fennel-repl-minify-code
- "minification is no longer preformed; variable value is ignored."
- "fennel-mode 0.7.0")
 
 (defun fennel-repl--input-filter (body)
   "Input filter for the comint process.